+++ ios/RNInstabug/InstabugBugReportingBridge.h
@@ -47,4 +47,8 @@
 
 - (void)setViewHierarchyEnabled:(BOOL)viewHirearchyEnabled;
 
+- (void)setDisclaimerText:(NSString *)text;
+
+- (void)setCommentMinimumCharacterCount:(NSNumber *)limit reportTypes:(NSArray *)reportTypes;
+
 @end